[runtime] Make Value::smi and Value::number handle all numeric types using new Numeric trait (#395)
## Summary

Refactor `Value::{smi, number}` and `Context::{smi, number}` to
generically handle all numeric types. This uses a new `Numeric` type
which we use to do compile-time dispatch into specialized
implementations for each numeric type.

- Replace all uses of `Value::from` for numeric types with
`Value::number`
- Use `Context::{smi, number}` everywhere handles are needed
- Remove all unnecessary casts on arguments to one of these functions 

## Tests

- All tests pass

+++ src/js/runtime/intrinsics/temporal/plain_date_time_constructor.rs
@@ -150,7 +150,7 @@ impl PlainDateTimeConstructor {
         let date_time_1 = to_temporal_date_time(cx, arg_1, NAME)?;
         let date_time_2 = to_temporal_date_time(cx, arg_2, NAME)?;
 
-        Ok(cx.smi(date_time_1.compare_iso(&date_time_2) as i32))
+        Ok(cx.smi(date_time_1.compare_iso(&date_time_2) as i8))
     }
 }
